+++ scenarioo-client/karma.conf.js
@@ -20,28 +20,55 @@
 module.exports = function (config) {
 
     config.set({
+        frameworks: ['jasmine'],
+
+
         // base path, that will be used to resolve files and exclude
         basePath: '',
 
         // list of files / patterns to load in the browser
         files: [
-            'app/components/angular/angular.js',
-            'app/components/angular-resource/angular-resource.js',
-            'app/components/angular-route/angular-route.js',
-            'app/components/angular-mocks/angular-mocks.js',
-            'app/components/angular-local-storage/dist/angular-local-storage.min.js',
-            'app/components/twigs/dist/twigs.js',
-            'app/components/svg.js/dist/svg.js',
-            'app/components/svg.draggable.js/dist/svg.draggable.js',
-            'app/components/svg-pan-zoom/dist/svg-pan-zoom.js',
-            'app/components/angular-unsavedChanges/dist/unsavedChanges.js',
+            'node_modules/angular/angular.js',
+            'node_modules/angular-mocks/angular-mocks.js',
             'app/*.js',
+            'app/*.ts',
             'app/!(components)/**/*.js',
             'test/mock/**/*.js',
-            'test/spec/**/*.js'
+            'test/mock/**/*.ts',
+            'test/spec/**/*.js',
+            'test/spec/**/*.ts'
         ],
 
-        frameworks: ['jasmine'],
+        preprocessors: {
+            'node_modules/angular/angular.js': ['webpack'],
+            'node_modules/angular-mocks/angular-mocks.js': ['webpack'],
+            'app/*.js': ['webpack'],
+            'app/*.ts': ['webpack'],
+            'app/!(components)/**/*.js': ['webpack'],
+            'app/!(components)/**/*.ts': ['webpack'],
+            'test/mock/**/*.js': ['webpack'],
+            'test/mock/**/*.ts': ['webpack'],
+            'test/spec/**/*.js': ['webpack'],
+            'test/spec/**/*.ts': ['webpack']
+        },
+
+        webpack: require('./configs/webpack.config.test'),
+
+        webpackMiddleware: {
+            noInfo: true,
+            // and use stats to turn off verbose output
+            stats: {
+                chunks: false
+            }
+        },
+
+        plugins: [
+            'karma-jasmine',
+            'karma-junit-reporter',
+            'karma-phantomjs-launcher',
+            'karma-webpack'
+        ],
+
 
         // list of files to exclude
         exclude: [],
@@ -49,11 +76,6 @@ module.exports = function (config) {
         // test results reporter to use
         // possible values: dots || progress || growl
         reporters: ['progress', 'junit'],
-        //reporters = ['progress', 'junit', 'coverage'];
-
-        //preprocessors = {
-        //    'app/scripts/services/*.js' : 'coverage'
-        //}
 
         // web server port
         port: 7070,
@@ -82,6 +104,7 @@ module.exports = function (config) {
         browserDisconnectTimeout : 10000,
         browserDisconnectTolerance : 1,
         browserNoActivityTimeout : 60000, //by default 10000
+
 
         // Continuous Integration mode
         // if true, it capture browsers, run tests and exit